Split actions into file & add edit action
